Housekeeping - remove service name and title
If the service name is a `const` in the plugin there is no reason not to include it directly in the layout. Adding it to the context means we have the option to override it but we can't foresee a time we ever would (famous last words!)

Conversely, we _always_ provide a page title from our controllers when returning a view. So, there seems little point in adding something we never use.