Move compile logic to its own function
What we had is taken directly from the example Vision provides for Nunjucks. But it depends on all the clever JavaScript stuff our conventions have us avoid.

So, this moves it to its own function where we also get a but more room to do our best to describe what is going on.