Move prepare() into its own method
Again, we just think it makes it clearer what we are actually assigning in the Vision's `options` object. Plus it gives us more space to explain what is going on.

We also do a bit of tidying up. The options we were providing default to those values so are unneeded and as we clearly set `relativeTo` in Vision's options a few lines above there is no need for protective coding to generate the path.