Handle unauthorized errors plus return safe codes (#472)
https://eaflood.atlassian.net/browse/WATER-4085
https://eaflood.atlassian.net/browse/WATER-4155

The final step of our pass-through authentication story is dealing with unauthorized requests. This is where Hapi will compare the scope defined in the credentials our `AuthPlugin` sets to what is in the route config. If they don't match Hapi will through an error with a [403 Forbidden](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/403) code.

When this happens our `ErrorsPlugin` will pick up the response and using `ErrorPagesService` determine if it should be redirected to an error page. Before this change, our logic would have deduced a [500 Internal Server Error](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500) and show the page for that.

Clearly, that's wrong but the solution is a little more complex. If you assume we don't render links to things a user cannot access, the request must have been manufactured i.e. directly entered into the browser. If we respond with a `403` and a nice 'You are not authorised' page we would be confirming the thing someone is trying to access exists, they just don't have the authority to. Similar to handling [user enumeration attacks](https://owasp.org/www-project-web-security-testing-guide/latest/4-Web_Application_Security_Testing/03-Identity_Management_Testing/04-Testing_for_Account_Enumeration_and_Guessable_User_Account) we don't want to do that. Instead, it is [recommended to return a 404](https://www.rfc-editor.org/rfc/rfc9110.html#name-403-forbidden).

So, we need to amend our error handling logic to check for `403` errors but ensure we respond with `404` instead.

Finally, whilst we are messing with the logic, we will also deal with the fact that returning a `500` means the user never sees our error page. This is because the service is behind a [F5 Silverline Managed Web Application Firewall (WAF)](https://aws.amazon.com/marketplace/pp/prodview-7ge77z55kayes) which blocks `5xx` responses and returns its own error page. (It does this to prevent apps from leaking info because of unhandled errors.)