Create links only is user authenticated
Thanks to [Add auth & user details to view context by default](https://github.com/DEFRA/water-abstraction-system/pull/477) we now have access to whether a user is authenticated in the global view context.

Using some Nunjucks _stuff_ (?? code, macros, filters - i dunno!) we can now use this to determine whether to generate nav links or not.

Note - Ordinarily we would so this kind of logic outside of the template. Our standard is to minimise logic in the view to the bare minimum, for example, if thing show x else show y.

But this is logic that is going to apply to _every_ page we display. So, it made sense to make an exception and include the logic in the layout template.