Update page header with user links (#474)
https://eaflood.atlassian.net/browse/WATER-4155

We are building our first proper page which will replace something the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) currently handles.

But before we get into the meat of it we need to update our header to include links specific to the user. These are

- **Change password**
- **Sign out**

> The legacy UI also displays **Contact information** but this is because it doesn't differentiate the internal and external sites. This link should only be shown to external users.

The complexity is we should only show them to authenticated users. So, in this change, we are going to use `/health/info` as a test! If an authenticated user happens to request the page we'll show the links. Else not.

Essentially, we're working out a way to do this that will hopefully work as we add more and more pages.