Add the nav bar template
The html for the layout and the SCSS all come from the [prototype](https://github.com/defra-design/wrls-prototype). We assume they are right!

One feature to note is the UI expects to highlight one of the nav bar options when on a page. The legacy service had it's own `ViewContext` plugin that created the global context from a variety of sources. Which nav bar item to highlight was one of them and they chose to set this in the config options of a route, along with the page title.

We think it will be clearer if data relevant to a page is route dependent, _all_ the data should come from the controller. For example

```javascript
  return h.view('info.njk', {
    pageTitle: 'Info',
    activeNavBar: 'bill-runs',
    ...pageData
  })
```

That way we know there are only 2 sources; the global context and the context we define in a controller.