Add phase banner to our page layout
https://eaflood.atlassian.net/browse/WATER-4155

We thought the [nav bar for internal users](https://github.com/DEFRA/water-abstraction-system/pull/478) would be the final piece. But we have since confirmed with our designer that until we know otherwise, we should also be including the [phase banner](https://design-system.service.gov.uk/components/phase-banner/) even though this is an internal app.

So, this change adds the banner to our layout. It will be displayed always, no matter if someone is authenticated or not.