Charge periods shown on bills are incorrect
https://eaflood.atlassian.net/browse/WATER-4179

It has been reported by a customer that that the charge period shown on a recent supplementary bill they received showed the whole financial year as a charge period, rather than the dates related to the charge versions and financial years.

The data passed to the charge module for supplementary appears to be using the financial year dates in all circumstances, when it should be using the charge period dates as shown on the transaction screen.

This PR will fix that issue by passing the correct dates to the CM.