Add new Multi-licence bill presenter (#496)
https://eaflood.atlassian.net/browse/WATER-4155

We are building our first proper page which will replace the bill summary page the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) currently provides. Rather than try to show all licences _and_ all their transactions for a bill on one page we'll instead list the licences. Users can then see the transaction details by selecting a licence.

We've already made several other changes to support this. This change adds a new presenter which takes details of the bill summaries for the multiple licences linked to it and formats them ready for use in a new view bill page.

---

We also spotted that we had missed a key field in our `FetchBillingAccountService` so rectified that in this change.