Add the template, SCSS and macro needed for page
Our first 'proper' template!

The intent going forward is that all presentation code should reside in the template. So, unless really necessary we should never see HTML or Nunjucks within our normal app code. But likewise the logic in our templates should just be about generating the layout not the content. So, `if` statements to see if something should be displayed, string concatenation to build links, or `for thing of things` in order to feed a govuk element are fine. Code to format a value however should be in a presenter.

Note - the SCSS content is styles that don't appear to exist in GOV.UK frontend package but do (in one version or another) exist in the prototype, the legacy UI, or both! We needed to add them in order to get the UI to match the prototype.