Add new view bill endpoint and template (#499)
https://eaflood.atlassian.net/browse/WATER-4155

We are building our first proper page which will replace the bill summary page the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) currently provides. Rather than try to show all licences _and_ all their transactions for a bill on one page we'll instead list the licences. Users can then see the transaction details by selecting a licence.

We've already made several other changes to support this. With the addition of [Add new view bill service](https://github.com/DEFRA/water-abstraction-system/pull/498) we can make the final change; creating the endpoint and adding the view template.

---

Our first 'proper' template!

The intent in the future is that all presentation code should reside in the template. So, unless necessary we should never see HTML or Nunjucks within our regular app code. But likewise, the logic in our templates should only be about generating the layout, not the content. So, `if` statements to see if something should be displayed, string concatenation to build links, or `for thing of things` to feed GOV.UK elements are fine. Code to format a value however should be in a presenter.

Note - the SCSS content is styles that don't appear to exist in GOV.UK frontend package but do (in one version or another) exist in the prototype, the legacy UI, or both! We needed to add them to get the UI to match the prototype.