Handle create 2PT bill run requests
https://eaflood.atlassian.net/browse/WATER-4204

We need to update the validation logic used that checks the incoming request to accept an optional start and end year for 2PT requests. We then need to pass this information down to our `app/services/billing/start-bill-run-process.service.js`.

The logic in that service will need to be updated to handle both 2PT and supplementary bill runs. For example, it assumes there will be multiple building periods calculated from the current year. For 2PT it will only be the current year, else the financial year provided in the request.