Linking SROC 2pt bill runs in UI to new review page
https://eaflood.atlassian.net/browse/WATER-4205

This pull request is preparing to create a new page for reviewing 2PT SROC bill runs. In the UI repo, changes have been made to redirect requests from links on the billing and batch list page that are related to 2PT SROC bill runs. These requests will now redirect to the new page in the system's repo.