Add new Minimum Charge Transaction presenter
https://eaflood.atlassian.net/browse/WATER-4156

Related to WATER-4155 and the work to replace the legacy bill page with our own we're now starting to build the page it will link to; a view of the bill licence and all its transactions.

This adds a presenter that given a `TransactionModel` instance whose charge type is `minimum_charge` will return an object containing formatted details about the transaction.

The presenter doesn't need to make any distinction between PRESROC and SROC transactions as they are both displayed in the same way.