Add new fetch bill licence service (#501)
https://eaflood.atlassian.net/browse/WATER-4156

Related to WATER-4155 and the work to replace the legacy bill page with our own we're now starting to build the page it will link to; a view of the bill licence and all its transactions.

This adds a service that will get the data needed for the page. Mainly that is the transaction detail but also information about the bill and licence needed for context on the page.

One thing to note is the service makes no distinction between SROC and PRESROC. However, some linked entities and fields are only used when a transaction is one or the other.