https://eaflood.atlassian.net/browse/WATER-4156
Related to WATER-4155 and the work to replace the legacy bill page with our own we're now starting to build the page it will link to; a view of the bill licence and all its transactions.

We added presenters for each of the transaction types

- [Minimum Charge Transaction presenter](https://github.com/DEFRA/water-abstraction-system/pull/505)
- [Compensation Charge Transaction presenter](https://github.com/DEFRA/water-abstraction-system/pull/506)
- [Standard Charge Transaction presenter](https://github.com/DEFRA/water-abstraction-system/pull/507)

Now we add a presenter that brings those together with details about the bill and bill run.