Add view bill licence service
https://eaflood.atlassian.net/browse/WATER-4156

Related to WATER-4155 and the work to replace the legacy bill page with our own we're now starting to build the page it will link to; a view of the bill licence and all its transactions.

We've already made several other changes to support this. This change adds a new service that orchestrates fetching the information needed ([Add new fetch bill licence service](https://github.com/DEFRA/water-abstraction-system/pull/501)), passing it through to a presenter ([Add View bill licence presenter](https://github.com/DEFRA/water-abstraction-system/pull/508)) and returning the result.