Add new templates
We could have had 1 template and used `{% if %}` blocks to determine what to show based on whether we are dealing with a presroc or sroc bill licence. But the page is already complex enough with trying to cram different content into different cells of a table without this extra headache.

So, in this case we've opted for separate templates and we leave the controller decide which view to use.