Add new view bill licence endpoint and template (#510)
https://eaflood.atlassian.net/browse/WATER-4156

Related to WATER-4155 and the work to replace the legacy bill page with our own we're now able to add the page it will link to; a view of the bill licence and all its transactions.

We've already made several other changes to support this. With the addition of [Add view bill licence service](https://github.com/DEFRA/water-abstraction-system/pull/509) we can make the final change; creating the endpoint and adding the view template.

---

Note about splitting the template

We could have had 1 template and used `{% if %}` blocks to determine what to show based on whether we are dealing with a PRESROC or SROC bill licence. But the page is already complex enough cramming different content into different cells of a table without this extra headache.

So, in this case, we've opted for separate templates and left the controller to decide which view to use.