Split MultiLicencePresenter into 2 new ones (#513)
https://eaflood.atlassian.net/browse/WATER-4132

In WATER-4155 and WATER-4156 we built 2 new pages for displaying a bill. WATER-4155 covered a new bill page that displayed details about the bill, the bill run it is in and its billing account. It then listed all licences connected to the bill.

Each licence if selected would take you to a new bill-licence page which displays the transactions for that licence in detail. This fixed the issue with the legacy code being unable to handle displaying licences with large (100s) numbers of licences.

But what if a bill has just 1 licence? In this case, the team agreed to show a table for 1 licence and requiring a user to click through for the details was superfluous.

So, WATER-4132 is about building a page for that scenario; a bill with just 1 licence. In this case, the information at the top of the page will remain the same. But instead of the licences table, we'll show the detailed breakdown of the transactions.

To support this work we need to split up the existing `MultiLicencePresenter` into 2 new presenters; `BillPresenter` and `LicenceSummaries`. We do this because we always need to format the template's bill and billing account data. But we only need to format licence summaries when dealing with a multi-licence bill.