Update view bill service
If we extended the way of testing the service we were currently using we'd need many hundreds of lines of fake data.

We have extensive tests of the presenters and fetch services already so this seemed unnecessary. So, instead we opted to switch to the approach used in test/services/bill-licences/view-bill-licence.service.test.js.