Add new single licence bill templates (#514)
https://eaflood.atlassian.net/browse/WATER-4132

In WATER-4155 and WATER-4156 we built 2 new pages for displaying a bill. WATER-4155 covered a new bill page that displayed details about the bill, the bill run it is in, and its billing account. It then listed all licences connected to the bill.

Each licence if selected would take you to a new bill-licence page (WATER-4156 ) which displays the transactions for that licence in detail. This fixed the issue (WATER-4070 ) with the legacy code being unable to display licences with large (100s) numbers of licences.

But what if a bill has just 1 licence? In this case, the team agreed that showing a table for 1 licence and requiring a user to click through for the details was superfluous.

So, WATER-4132 is about building a page for that scenario; a bill with just 1 licence. In this change, we add the new templates needed to do this along with changes to the controller and ViewBillService to determine which template to show.