Rename Invoice Account to Billing Account (#515)
https://github.com/DEFRA/water-abstraction-team/issues/97

We overlooked this in [the Great Rename](https://github.com/DEFRA/water-abstraction-team/issues/97) because it isn't in the `water` schema. In the UI and amongst the team what the DB calls 'Invoice Account' is referred to as 'Billing Account'. In our latest changes which focused on building new bill run pages, we had to reference this table and model a lot, which for anyone new is going to look confusing. For example, see `app/services/bills/fetch-billing-account.service.js`.

```javascript
/**
 * Fetches the billing account and associated details for a given invoice account ID
 * @module FetchBillingAccountService
 */

const InvoiceAccount = require('../../models/crm-v2/invoice-account.model.js')

```

So, we're sorting that out in this change, renaming the invoice account model to `BillingAccountModel`.