Rename services/billing to bill-runs (#516)
It was noted during a pairing session that the name `app/services/billing` is inconsistent with our current convention.

For example, we have a `/bills` endpoint determined by `bills.routes.js`. Requests to that endpoint are handled by `bills.controller.js` which in turn refers to services in `app/services/bills`.

The `/bill-runs` endpoint follows this convention until you get to the services folder. Then everything it references is found in `app/services/billing`.

We know why this is; handling SROC supplementary billing was our first major feature on this service and _everything_ seemed to be called **BillingSomething**!

But as we are in the process of delivery our next major billing feature (two-part tariff bill runs) now seems a good time to correct the folder name and make it consistent.