Fix Airbrake when behind a proxy
https://github.com/DEFRA/water-abstraction-team/issues/104

We copied the airbrake plugin we have from [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) where it works great. It also works great locally and we just assumed **water-abstraction-system** wasn't firing any errors in production. But we've realised no errors are getting through in our AWS environments because the apps sit behind a proxy and we haven't configured Airbrake to handle this!

In this change we follow the Airbrake docs and properly configure Airbrake (and yes - we have tested it this time!)

---

According to the [Airbrake docs](https://github.com/airbrake/airbrake-js/tree/master/packages/node#nodejs-request-and-proxy) the only way to do this is with the following.

> To use the [request](https://github.com/request/request) HTTP client, pass the request option which accepts a request wrapper:
>
> ```javascript
> const airbrake = new Notifier({
>   // ...
>   request: request.defaults({ proxy: 'http://localproxy.com' }),
> });
>

It's unfortunate that Airbrake is relying on [request](https://github.com/request/request) as it has been deprecated since Feb 2020. But this directly impacts us because the only solution we've found is to also add **request** as a dependency. Normally, we've been able to `require()` sub-dependencies. But something about the way the Airbrake package is built means that's not the case here.