Fix display of credit transactions in bill views
https://eaflood.atlassian.net/browse/WATER-4131

In testing we've spotted that the display of credit values in certain screens has gone awry!

In most cases we'll display `£150.25 credit` which is correct. But there are times, mainly in tables, where we need to show the signed version, for example `-£150.25`. A change to clean up our money formatters in `BasePresenter` broke this so now negative values in the licence summaries table and in the transaction tables total row are always being displayed as positive.

This change updates the formatter to allow callers to specify whether a sign should be returned. It also fixes all the JSDoc comments we forgot to update when we made the change. Doh!