Remove check of source for bill licence page
We took the need to check for both the batch type being supplementary and the source being NALD from the legacy code. It's why even in `production` you'll see some bills display without the credits column.

But in testing nothing seems to have blown up when we looked at bill licences connected to `nald` supplementary bill runs. So, we're removing the check from our logic.